Jquery routing for toybox links
I might have a solution to the viewport problem,
just using divs that are hiding on the outside
of the page which pop in and push the center elements
out of the way. Should be easy to implement because
the page is so simple. Also, plan is to animate the
svg gradient to make up for how the above idea looks
less cool than what I'd had in mind.