Fixed svg icons popping in ie and firefox
Just discovered when doing cross-browser tests that
lines in the svg icons were popping in and out, like
the browsers were having a hard time rendering them.
It must have been an arifact of the way I was irresponsibly
scaling the icons down from like 600px to 144 - chrome seemed
to handle that fine, but ie and firefox like it much better
now that the svgs themselves are at 144x144px. Now I know!