Commit before branching
Going to experiment with finding a different way to
center the centerpiece thing - it's currently not allowing
me to move the damn thing around due to the fact that its
position is relative and the flexbox is in a column. Going
to mess around in the branch til i find somehting that works.