Icons moving, but at what cost?
Really weird thing with inconsisent left/right shifting.
Going to delete styling and copy in the left popin box
until I figure out what it is. Also, mobile layout
still huge issue, it's beginning to do the thing I want
(line up in a row on the sides) but it's going to take
a lot of finnagling. Is that how you spell that?