Discover the magic of weinre + the hubris of man
Weinre is amazing and allowed me to debug on my phone, which allowed me
to realize that I now have a lot of problems. ios safari
on my iphone definitely doesn't work even a little bit,
the translate animations for all the boxes aren't doing
anything, and neither is the opacity changing on the
centerpiece. I'm about to ~learn so much~ I can just tell