A new and serious layout issue appeared!
iOS safari text sizing doesn't agree with my responsive
layout! At all! Right now, I've put a -webkit-text-size-adjust:100%
on the body, which appears to be not the best idea according to the
internet and general standards of accesibiltity. I'll have to do more
research re why safari is doing this with my layout.