achieve enlightenment re: mobile scrolling issue
It's because I put hammerjs to listen for swipe events on
touch events - for example, scrolling! - on the element it's
being used on, which makes sense even though it's been the cause
of so much sadness visited upon the world.

Now I need to find a way to get hammerjs to play nicely with scrolling...
inertia scrolling without too much of a processing cost would be
great, but maybe I don't need to leverage an honestToGod js library
for that. This is good tho! I understand the issue! Mobile responsiveness
will not defeat me this day