Achieve understanding of misplaced svgs issue
It defo has to do with the way the page loads content - which explains
the differences cross-browser. I would have hoped that doc.ready would
avoid that kind of thing, but oh well. I tested the thing by redrawing the
svgs using a manual call to the drawSVG function, which I also stuck into
a global namespace variable. Please remember to remove that from your code
if you don't need to call the drawSVG function from the dom anywhere (which
you probably wont!)