Achieve yet deeper understanding of stupid svg lines
It's definitely the layout changing as a result of the lines
being drawn, at least that's what it is in firefox and ie. I'm
guessing a pos: relative parent div is being scaled upward by
the inclusion of the svg object. Most likely it's an issue that
can be solved by fiddling with where I stick the svg container
in the html. I'll fiddle with that later, I'm pretty over these
svg lines haha.

Right now going to hop over to illustrator and finish the centerpiece
icons and loading.svg (lol), which I've been putting off for no
good reason for like weeks haha