Responsive layout and style for contact-box
Still need to look into submitting the actual form, so maybe
the site won't be static after all. There are worse problems.
One of those problems is touch event handling, I need to resolve
that issue as soon as possible... but after that, and after whatever
other little style cleanups I need to do, we should be actually
honest-to-god ready to launch. Won't that feel amazing.