Workaround for touch events on icons in iOS
The issue was that, while pointer-events:none was doing a good
job in the desktop of allowing click events to bubble out of the svg
inside the button, it didn't seem to be working in iOS. I tried
writing a handler to manually bubble up the event to the parent (the button)
but it didn't work and I didn't bring my iphone usb cable with me so it's hard
to debug... Anyway, I just stuck all the svgs on the z-index right below
so that the button is technically on top. Hacky, but it works in this case, thank
god the button doesn't have a background.