Commit before adding tab sliders to boxes
(the user needs to be prompted to swipe in a direction to close
the boxes on mobile, obvs. So I'm going to add them, by using
a wrapper div with a high z-index tab on the side facing the center.
It will hover above the inner div with the scrolling etc)