Prepare server for launch
I'm waiting on like two lbs of ribs to finish cooking, they were on
sale at the grocery store(!). I had to do a lot of finnagling with gulp,
but everything is getting ported to the server/dist folder properly.
The cat I live with kept coming up to me and headbutting my laptop
as I smh-ed at myself for forgetting to build the dist folder before
trying to launch the app. I got too used to autoreload. It's been
storming ever since I left the library. It's a good day to launch
the website.