Refactor svg line drawing code
Lots of changes:
	- svg line drawing abstracted out to its own module
	- client-app updated to use the module instead of hardcoding
		the svg line drawing into the jQuery.each call that
		iterated through the about-skills-items. Svg line highlighting
		function also works like normal.
	- (minor) fixed the annoying line coordinate issue using jQuery's outerHeight
	- index.html updated to incorporate new script
	- main.css updated to reference svg element via ID rather than class

TODO: add sourcemaps so that everyone can see my beautiful (or really,
just less ugly) code. Also see if there are any sourcemap-generating
complications resulting from useref concatenating everything on build.

TODO: Refactor the rest of it! But I think the svgs were the jumbliest part.
Not to mention the part that's most useful for other projects...